+++ be/src/vec/data_types/serde/data_type_hll_serde.cpp
@@ -65,7 +65,7 @@ Status DataTypeHLLSerDe::serialize_one_cell_to_json(const IColumn& column, int64
 
 Status DataTypeHLLSerDe::deserialize_column_from_json_vector(IColumn& column,
                                                              std::vector<Slice>& slices,
-                                                             int* num_deserialized,
+                                                             uint64_t* num_deserialized,
                                                              const FormatOptions& options) const {
     DESERIALIZE_COLUMN_FROM_JSON_VECTOR();
     return Status::OK();
